+++ src/IntervalTest.php
@@ -938,14 +938,14 @@ final class IntervalTest extends TestCase
             $interval->toNotation(IntervalNotation::Iso8601StartEnd),
             $interval->toNotation(IntervalNotation::Iso8601StartEnd, Unit::Minute)
         );
-        self::assertSame('10:00:00/12:00:00', $interval->truncateTo(Unit::Second)->toNotation(IntervalNotation::Iso8601StartEnd));
-        self::assertSame('10:00:01/12:00:00', $interval->roundTo(Unit::Second)->toNotation(IntervalNotation::Iso8601StartEnd));
-        self::assertSame('[600,720[', $interval->truncateTo(Unit::Second)->toNotation(IntervalNotation::Bourbaki, Unit::Minute));
-        self::assertSame('[600.016667,720[', $interval->roundTo(Unit::Second)->toNotation(IntervalNotation::Bourbaki, Unit::Minute));
+        self::assertSame('10:00:00/12:00:00', $interval->roundTo(Unit::Second, RoundingMode::Floor)->toNotation(IntervalNotation::Iso8601StartEnd));
+        self::assertSame('10:00:01/12:00:00', $interval->roundTo(Unit::Second, RoundingMode::Round)->toNotation(IntervalNotation::Iso8601StartEnd));
+        self::assertSame('[600,720[', $interval->roundTo(Unit::Second, RoundingMode::Floor)->toNotation(IntervalNotation::Bourbaki, Unit::Minute));
+        self::assertSame('[600.016667,720[', $interval->roundTo(Unit::Second, RoundingMode::Round)->toNotation(IntervalNotation::Bourbaki, Unit::Minute));
         self::assertSame('[600.016667,720[', Interval::fromNotation(
-                '[600.016667,720[',
-                IntervalNotation::Bourbaki,
-                Unit::Minute
-            )->toNotation(IntervalNotation::Bourbaki, Unit::Minute));
+            '[600.016667,720[',
+            IntervalNotation::Bourbaki,
+            Unit::Minute
+        )->toNotation(IntervalNotation::Bourbaki, Unit::Minute));
     }
 }