+++ src/RoundingMode.php
@@ -1,0 +1,12 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Bakame\Tokei;
+
+enum RoundingMode
+{
+    case Floor;
+    case Round;
+    case Ceil;
+}